add handling exception for Employees who has animals

+++ src/main/java/pl/schronisko/service/UserService.java
@@ -2,11 +2,14 @@ package pl.schronisko.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import pl.schronisko.exception.UserHasAnimalException;
 import pl.schronisko.exception.UserNotFoundException;
+import pl.schronisko.model.Animal;
 import pl.schronisko.model.User;
 import pl.schronisko.repository.UserRepository;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -14,6 +17,8 @@ public class UserService {
 
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private AnimalService animalService;
 
     public List<User> listAll() {
         return (List<User>) userRepository.findAll();
@@ -26,10 +31,13 @@ public class UserService {
         if (result.isPresent()) return result.get();
         throw new UserNotFoundException("Could not find any user with ID: "+id);
     }
-    public void delete(Integer id) throws UserNotFoundException {
+    public void delete(Integer id) throws UserNotFoundException, UserHasAnimalException {
         Long count = userRepository.countById(id);
         if(count == null || count == 0) {
             throw new UserNotFoundException("Could not find any user with ID: "+id);
+        }
+        if(hasUserAnyAnimal(id)) {
+            throw new UserHasAnimalException("User with ID: "+ id + "has animal.");
         }
         userRepository.deleteById(id);
     }
@@ -40,5 +48,16 @@ public class UserService {
     public User findUserByEmail(String email) {
         return userRepository.findByEmail(email);
     }
+
+    public boolean hasUserAnyAnimal(Integer idUser) {
+        List<Animal> animals = animalService.listAll();
+        for(Animal animal : animals) {
+            if(Objects.equals(animal.getIdUser().getId(), idUser)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
 
 }
